+++ modules/tinymce/src/plugins/fullscreen/main/ts/api/Events.ts
@@ -2,6 +2,7 @@ import Editor from 'tinymce/core/api/Editor';
 
 const fireFullscreenStateChanged = (editor: Editor, state: boolean): void => {
   editor.dispatch('FullscreenStateChanged', { state });
+  editor.dispatch('ResizeEditor', { state });
 };
 
 export {